Rename: utils, components 분리
src에 있던 파일들을 utils와 components로 분리
