--- src/Location.js
@@ -1,7 +1,0 @@
-export default function Location({
-  $target,
-  initialState
-}) {
-  
-
-}
