+++ src/components/Location.js
@@ -1,0 +1,7 @@
+// export default function Location({
+//   $target,
+//   initialState
+// }) {
+  
+
+// }
