+++ src/main.js
@@ -1,4 +1,4 @@
-import { request } from "./api.js"
+
 import App from "./App.js"
